update env for some little fixes; python 3 ncol div fix, corelp_xml import

+++ corpkit/plotter.py
@@ -608,7 +608,7 @@ def plotter(df,
     if legend:
         if num_to_plot > 6:
             if not kwargs.get('ncol'):
-                kwargs['ncol'] = num_to_plot / 7
+                kwargs['ncol'] = int(num_to_plot / 7)
         # kwarg options go in leg_options
         leg_options = {'framealpha': leg_alpha,
                        'shadow': kwargs.get('shadow', False),
